+++ src/shapepy/bool2d/__init__.py
@@ -3,8 +3,8 @@ Defines the internal module that computes the boolean
 operations between subsets
 """
 
+from .base import Future
 from .boolean import intersect, unite
-from .shape import Future
 
 Future.unite = unite
 Future.intersect = intersect
